Update to use newer tools.
In particular, the old version of mCRL2 did not work anymore with recent
compilers.

If you want to use generated parity games and be sure to have the games
identical to those generated by the previous tool version, use the games
for download in README.md for the time being.

I still need to check whether there are any changes in the generated
parity games now that we use a different mCRL2 version.